test: separate custom assert into multiple class
