+++ core/instrumentationtest/src/main/kotlin/com/waffiq/bazz_movies/core/instrumentationtest/Helper.kt
@@ -1,20 +1,14 @@
 package com.waffiq.bazz_movies.core.instrumentationtest
 
-import android.text.method.PasswordTransformationMethod
 import android.view.View
-import android.widget.EditText
 import androidx.lifecycle.Lifecycle
 import androidx.test.core.app.ActivityScenario
 import androidx.test.espresso.Espresso.onView
-import androidx.test.espresso.UiController
-import androidx.test.espresso.ViewAction
 import androidx.test.espresso.ViewAssertion
 import androidx.test.espresso.assertion.ViewAssertions.matches
-import androidx.test.espresso.matcher.BoundedMatcher
 import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
 import androidx.test.espresso.matcher.ViewMatchers.isRoot
-import androidx.test.espresso.matcher.ViewMatchers.withId
-import org.hamcrest.Description
+import com.waffiq.bazz_movies.core.instrumentationtest.CustomAssertions.waitFor
 import org.hamcrest.Matcher
 
 object Helper {
@@ -42,14 +36,6 @@ object Helper {
       Thread.sleep(DELAY_TIME)
     }
   }
-
-  fun waitFor(millis: Long = 1000): ViewAction =
-    object : ViewAction {
-      override fun getConstraints(): Matcher<View> = isRoot()
-      override fun getDescription(): String = "Wait for $millis milliseconds"
-      override fun perform(uiController: UiController, view: View) =
-        uiController.loopMainThreadForAtLeast(millis)
-    }
 
   fun waitUntil(matcher: Matcher<View>, timeout: Long = 5000): ViewAssertion {
     return ViewAssertion { view, _ ->
@@ -82,45 +68,5 @@ object Helper {
       }
     }
     onView(matcher).check(matches(isDisplayed()))
-  }
-
-  fun isPasswordHidden(): Matcher<View> =
-    object : BoundedMatcher<View, EditText>(EditText::class.java) {
-      override fun describeTo(description: Description) {
-        description.appendText("Password is hidden")
-      }
-
-      // check if password is hidden
-      override fun matchesSafely(editText: EditText): Boolean =
-        editText.transformationMethod is PasswordTransformationMethod
-    }
-
-  fun withErrorText(expectedError: String): Matcher<View> =
-    object : BoundedMatcher<View, EditText>(EditText::class.java) {
-      override fun describeTo(description: Description) {
-        description.appendText("with error text: $expectedError")
-      }
-
-      override fun matchesSafely(editText: EditText): Boolean =
-        editText.error?.toString() == expectedError
-    }
-
-  private fun withoutError(): Matcher<View> =
-    object : BoundedMatcher<View, EditText>(EditText::class.java) {
-      override fun describeTo(description: Description) {
-        description.appendText("without error text")
-      }
-
-      override fun matchesSafely(editText: EditText): Boolean = editText.error == null
-    }
-
-  fun Int.hasErrorText(text: String) {
-    onView(withId(this))
-      .check(matches(withErrorText(text)))
-  }
-
-  fun Int.hasNoError() {
-    onView(withId(this))
-      .check(matches(withoutError()))
   }
 }
