+++ feature/about/src/androidTest/kotlin/com/waffiq/bazz_movies/feature/about/ui/AboutActivityTest.kt
@@ -10,9 +10,9 @@ import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.filters.LargeTest
 import com.waffiq.bazz_movies.core.common.utils.Constants.BAZZ_MOVIES_LINK
 import com.waffiq.bazz_movies.core.common.utils.Constants.TMDB_LINK_MAIN
-import com.waffiq.bazz_movies.core.instrumentationtest.ViewMatcher.isClickable
-import com.waffiq.bazz_movies.core.instrumentationtest.ViewMatcher.isDisplayed
-import com.waffiq.bazz_movies.core.instrumentationtest.ViewMatcher.performClick
+import com.waffiq.bazz_movies.core.instrumentationtest.CustomViewActions.performClick
+import com.waffiq.bazz_movies.core.instrumentationtest.CustomViewMatchers.isClickable
+import com.waffiq.bazz_movies.core.instrumentationtest.CustomViewMatchers.isDisplayed
 import com.waffiq.bazz_movies.feature.about.R.id.btn_about_us
 import com.waffiq.bazz_movies.feature.about.R.id.iv_tmdb_logo
 import com.waffiq.bazz_movies.feature.about.R.id.tv_about_text
