+++ feature/detail/src/androidTest/kotlin/com/waffiq/bazz_movies/feature/detail/ui/MediaDetailActivityInteractionTest.kt
@@ -7,15 +7,15 @@ import com.waffiq.bazz_movies.core.common.utils.Constants.NAN
 import com.waffiq.bazz_movies.core.common.utils.Constants.TV_MEDIA_TYPE
 import com.waffiq.bazz_movies.core.designsystem.R.string.cancel
 import com.waffiq.bazz_movies.core.designsystem.R.string.submit
-import com.waffiq.bazz_movies.core.instrumentationtest.ViewMatcher.clickItemAt
-import com.waffiq.bazz_movies.core.instrumentationtest.ViewMatcher.doesHaveText
-import com.waffiq.bazz_movies.core.instrumentationtest.ViewMatcher.doesNotHaveText
-import com.waffiq.bazz_movies.core.instrumentationtest.ViewMatcher.performAction
-import com.waffiq.bazz_movies.core.instrumentationtest.ViewMatcher.performClick
-import com.waffiq.bazz_movies.core.instrumentationtest.ViewMatcher.performScrollTo
-import com.waffiq.bazz_movies.core.instrumentationtest.ViewMatcher.performSwipeDown
-import com.waffiq.bazz_movies.core.instrumentationtest.ViewMatcher.performSwipeLeft
-import com.waffiq.bazz_movies.core.instrumentationtest.ViewMatcher.performTextClick
+import com.waffiq.bazz_movies.core.instrumentationtest.CustomRecyclerViewActions.clickItemAt
+import com.waffiq.bazz_movies.core.instrumentationtest.CustomViewActions.performAction
+import com.waffiq.bazz_movies.core.instrumentationtest.CustomViewActions.performClick
+import com.waffiq.bazz_movies.core.instrumentationtest.CustomViewActions.performScrollTo
+import com.waffiq.bazz_movies.core.instrumentationtest.CustomViewActions.performSwipeDown
+import com.waffiq.bazz_movies.core.instrumentationtest.CustomViewActions.performSwipeLeft
+import com.waffiq.bazz_movies.core.instrumentationtest.CustomViewActions.performTextClick
+import com.waffiq.bazz_movies.core.instrumentationtest.CustomViewMatchers.doesHaveText
+import com.waffiq.bazz_movies.core.instrumentationtest.CustomViewMatchers.doesNotHaveText
 import com.waffiq.bazz_movies.feature.detail.R.id.btn_back
 import com.waffiq.bazz_movies.feature.detail.R.id.btn_favorite
 import com.waffiq.bazz_movies.feature.detail.R.id.btn_more_recommendation
