+++ feature/list/src/androidTest/kotlin/com/waffiq/bazz_movies/feature/list/ListActivityTest.kt
@@ -6,12 +6,12 @@ import androidx.paging.PagingData
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.espresso.intent.Intents
 import com.waffiq.bazz_movies.core.designsystem.R.id.btn_try_again
-import com.waffiq.bazz_movies.core.instrumentationtest.ViewMatcher.doesNotExist
-import com.waffiq.bazz_movies.core.instrumentationtest.ViewMatcher.isDisplayed
-import com.waffiq.bazz_movies.core.instrumentationtest.ViewMatcher.isNotDisplayed
-import com.waffiq.bazz_movies.core.instrumentationtest.ViewMatcher.isVisible
-import com.waffiq.bazz_movies.core.instrumentationtest.ViewMatcher.performClick
-import com.waffiq.bazz_movies.core.instrumentationtest.ViewMatcher.performSwipeDown
+import com.waffiq.bazz_movies.core.instrumentationtest.CustomViewActions.performClick
+import com.waffiq.bazz_movies.core.instrumentationtest.CustomViewActions.performSwipeDown
+import com.waffiq.bazz_movies.core.instrumentationtest.CustomViewMatchers.doesNotExist
+import com.waffiq.bazz_movies.core.instrumentationtest.CustomViewMatchers.isDisplayed
+import com.waffiq.bazz_movies.core.instrumentationtest.CustomViewMatchers.isNotDisplayed
+import com.waffiq.bazz_movies.core.instrumentationtest.CustomVisibilityMatchers.isVisible
 import com.waffiq.bazz_movies.core.uihelper.state.UIState
 import com.waffiq.bazz_movies.feature.list.R.id.btn_close
 import com.waffiq.bazz_movies.feature.list.R.id.btn_toggle_layout
