+++ feature/login/src/androidTest/kotlin/com/waffiq/bazz_movies/feature/login/ui/LoginActivityTest.kt
@@ -23,16 +23,16 @@ import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import com.waffiq.bazz_movies.core.designsystem.R.string.guest_user
 import com.waffiq.bazz_movies.core.domain.UserModel
-import com.waffiq.bazz_movies.core.instrumentationtest.Helper.hasErrorText
-import com.waffiq.bazz_movies.core.instrumentationtest.Helper.hasNoError
-import com.waffiq.bazz_movies.core.instrumentationtest.Helper.isPasswordHidden
-import com.waffiq.bazz_movies.core.instrumentationtest.ViewMatcher.doesHaveText
-import com.waffiq.bazz_movies.core.instrumentationtest.ViewMatcher.isDisplayed
-import com.waffiq.bazz_movies.core.instrumentationtest.ViewMatcher.isNotDisplayed
-import com.waffiq.bazz_movies.core.instrumentationtest.ViewMatcher.isNotEnable
-import com.waffiq.bazz_movies.core.instrumentationtest.ViewMatcher.performClick
-import com.waffiq.bazz_movies.core.instrumentationtest.ViewMatcher.performType
-import com.waffiq.bazz_movies.core.instrumentationtest.ViewMatcher.withDrawable
+import com.waffiq.bazz_movies.core.instrumentationtest.CustomAssertions.isPasswordHidden
+import com.waffiq.bazz_movies.core.instrumentationtest.CustomViewActions.performClick
+import com.waffiq.bazz_movies.core.instrumentationtest.CustomViewActions.performType
+import com.waffiq.bazz_movies.core.instrumentationtest.CustomViewMatchers.doesHaveText
+import com.waffiq.bazz_movies.core.instrumentationtest.CustomViewMatchers.hasErrorText
+import com.waffiq.bazz_movies.core.instrumentationtest.CustomViewMatchers.hasNoError
+import com.waffiq.bazz_movies.core.instrumentationtest.CustomViewMatchers.isDisplayed
+import com.waffiq.bazz_movies.core.instrumentationtest.CustomViewMatchers.isNotDisplayed
+import com.waffiq.bazz_movies.core.instrumentationtest.CustomViewMatchers.isNotEnable
+import com.waffiq.bazz_movies.core.instrumentationtest.CustomViewMatchers.withDrawable
 import com.waffiq.bazz_movies.core.user.ui.viewmodel.UserPreferenceViewModel
 import com.waffiq.bazz_movies.feature.login.R.drawable.ic_eye_off
 import com.waffiq.bazz_movies.feature.login.R.id.activity_login
